+++ backend/source/src/users/users.controller.ts
@@ -1,12 +1,16 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from "@nestjs/common";
+import { Body, Controller, Delete, Get, HttpException, HttpStatus, Param, Post, Put } from "@nestjs/common";
 import { UsersService } from "./users.service";
 import { Users } from "./users.schema";
 import { ApiResponse, ApiTags } from "@nestjs/swagger";
+import { CartsService } from "src/carts/carts.service";
 
 @ApiTags('users')
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) { }
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly cartsService: CartsService
+  ) { }
 
   @Post()
   @ApiResponse({
@@ -45,6 +49,16 @@ export class UsersController {
     type: Users
   })
   async delete(@Param('id') id: string) {
+    let toDeleteUser = await this.usersService.readById(id)
+
+    let relatedCarts = await this.cartsService.readAll({
+      email: toDeleteUser.email
+    })
+
+    if (relatedCarts.length != 0 && relatedCarts[0].items.length != 0) {
+      throw new HttpException("This user has some products in his cart", HttpStatus.FORBIDDEN)
+    }
+
     return await this.usersService.delete(id);
   }
-}
+}
